c3 lesson 4 completed and some code restrucutre into separate .go files
